läs boken
det står i boken

+++ app.rb
@@ -4,7 +4,9 @@ require_relative 'module.rb'
 
 include Database
 
-enable :sessions
+	use Rack::Session::Cookie,	:key => 'rack.session',
+								:expire_after => 62312738213721837897,
+								:secret => 'myhiddensecret'
 
 	get '/' do
 		x = "b" + rand(1..3).to_s + "box"
@@ -37,25 +39,27 @@ enable :sessions
 		end
 		message_limit = 10
 		group_id = params["id"]
-		users = fetch_users_from_group(group_id)
+		users = fetch_userinfo_from_group(group_id)
 		messages = fetch("*", "messages", "groupid", group_id)
-		leader_id = fetch("groupleaderid", "groups", "id", group_id).join
-		logged_in_user_id = fetch_userinfo(session[:user], "id").join
-		usernames = []
-		users.each do |user_id|
-			usernames << fetch("name", "users", "id", user_id).join
-		end
-		user_id_name = {}
-		users.each do |user_id|
-			user_id_name[user_id.join] = fetch("name", "users", "id", user_id.join).join
-		end
-		while messages.length > message_limit
-			messages.delete_at(0)
-		end
-		if leader_id == logged_in_user_id
-			slim(:groupifleader, locals:{usernames:usernames, messages:messages, group_id:group_id, user_id_name:user_id_name})
+		leader_id = fetch_group_leader(group_id).join
+		logged_in_user = fetch_userinfo(session[:user], "")
+		logged_in_user[0].delete_at(2)
+		if users.include?(logged_in_user[0])
+			while messages.length > message_limit
+				messages.delete_at(0)
+			end
+			if leader_id.to_s == logged_in_user[0][0].to_s
+				all_users = fetch_all_users()
+				all_users = all_users.reject {|w| users.include? w}
+				slim(:groupifleader, locals:{users:users, messages:messages, group_id:group_id, all_users:all_users})
+			else
+				slim(:group, locals:{users:users, messages:messages, group_id:group_id})
+			end
 		else
-			slim(:group, locals:{usernames:usernames, messages:messages, group_id:group_id, user_id_name:user_id_name})
+			session[:fail_message] = "You are not allowed to do that"
+			session[:user] = nil
+			session[:redirect_to] = "/logout"
+			redirect('/fail')
 		end
 	end
 
@@ -97,18 +101,23 @@ enable :sessions
 		username = params["username"]
 		password2 = params["password2"]
 		password = params["password"]
-		if username.length <= 0
-			session[:fail_message] = "Username too short"
+		if username.length < 3 || username.length > 20
+			session[:fail_message] = "Username must be between 3 and 20 characters"
 			session[:redirect_to] = "/register"
 			redirect('/fail')
 		end
-		if password.length <= 0
-			session[:fail_message] = "Password too short"
+		if password.length < 3 || password.length > 40
+			session[:fail_message] = "Password must be between 3 and 40 characters"
 			session[:redirect_to] = "/register"
 			redirect('/fail')
 		end
 		if password2 != password
 			session[:fail_message] = "Passwords does not match"
+			session[:redirect_to] = "/register"
+			redirect('/fail')
+		end
+		if username.strip == ""
+			session[:fail_message] = "Username must contain letters"
 			session[:redirect_to] = "/register"
 			redirect('/fail')
 		end
@@ -127,6 +136,17 @@ enable :sessions
 	post '/start/groups/create' do
 		db = connect
 		group_name = params["group_name"]
+		if group_name.length < 3 || group_name.length > 60
+			session[:fail_message] = "Group name must be between 3 and 60 characters"
+			session[:redirect_to] = "/start"
+			redirect('/fail')
+		end
+		p group_name
+		if group_name.strip == ""
+			session[:fail_message] = "Group name must contain letters"
+			session[:redirect_to] = "/start"
+			redirect('/fail')
+		end
 		user_id = fetch_userinfo(session[:user], "id").join
 		db.execute("INSERT INTO groups (name, groupleaderid) VALUES (?,?)",[group_name,user_id])
 		group_id = fetch("id", "groups", "name", group_name).join
@@ -139,21 +159,30 @@ enable :sessions
 			redirect('/')
 		end
 		group_id = params["group_id"]
-		users = fetch_users_from_group(group_id)
+		users = fetch_userinfo_from_group(group_id)
 		message = params["message"]
-		user_id = fetch_userinfo(session[:user], "id").join
-		checker = 0
-		users.each do |user|
-			if user[0].to_s == user_id
-				checker = 1
-			end
+		if message.length < 1 || message.length > 40
+			session[:fail_message] = "Message must be between 1 and 40 characters"
+			session[:redirect_to] = "/start/groups/#{group_id}"
+			redirect('/fail')
 		end
-		if checker == 0
-			redirect('/logout')
+		if message.strip == ""
+			session[:fail_message] = "Message must contain letters"
+			session[:redirect_to] = "/start/groups/#{group_id}"
+			redirect('/fail')
 		end
-		db = connect
-		db.execute("INSERT INTO messages (userid, groupid, message) VALUES (?,?,?)",[user_id,group_id,message])
-		redirect("/start/groups/#{group_id}")
+		logged_in_user = fetch_userinfo(session[:user], "")
+		logged_in_user[0].delete_at(2)
+		if users.include?(logged_in_user[0])
+			db = connect
+			db.execute("INSERT INTO messages (userid, groupid, message) VALUES (?,?,?)",[logged_in_user[0][0],group_id,message])
+			redirect("/start/groups/#{group_id}")
+		else
+			session[:fail_message] = "You are not allowed to do that"
+			session[:user] = nil
+			session[:redirect_to] = "/logout"
+			redirect('/fail')
+		end
 	end
 
 end
